feat: extracts relationship input for exteral use
