+++ packages/ui/src/fields/Relationship/types.ts
@@ -1,5 +1,12 @@
 import type { I18nClient } from '@payloadcms/translations'
-import type { ClientCollectionConfig, ClientConfig, FilterOptionsResult } from 'payload'
+import type {
+  ClientCollectionConfig,
+  ClientConfig,
+  CollectionSlug,
+  FilterOptionsResult,
+  StaticDescription,
+  StaticLabel,
+} from 'payload'
 
 export type Option = {
   allowEdit: boolean
@@ -63,3 +70,44 @@ export type GetResults = (args: {
   sort?: boolean
   value?: Value | Value[]
 }) => Promise<void>
+
+export type RelationshipInputProps = {
+  readonly AfterInput: React.ReactNode // optional
+  readonly allowCreate: boolean // optional
+  readonly allowEdit: boolean // optional
+  readonly appearance: 'drawer' | 'select' // optional
+  readonly BeforeInput: React.ReactNode // optional
+  readonly className: string // optional
+  readonly Description: React.ReactNode // optional
+  readonly description: StaticDescription // optional
+  readonly Error: React.ReactNode // optional
+  readonly filterOptions: FilterOptionsResult // optional
+  readonly initialValue: Value | Value[] // optional
+  readonly isPolymorphic: boolean
+  readonly isSortable: boolean // optional
+  readonly Label: React.ReactNode // optional
+  readonly label: StaticLabel // optional
+  readonly localized: boolean // optional
+  readonly maxResultsPerRequest: number // optional
+  readonly maxRows: number // optional
+  readonly minRows: number // optional
+  readonly path: string
+  readonly readOnly: boolean // optional
+  readonly relationTo: string[]
+  readonly required: boolean // optional
+  readonly showError: boolean // optional
+  readonly sortOptions: Partial<Record<CollectionSlug, string>> // optional
+  readonly style: React.CSSProperties // optional
+} & SharedRelationshipInputProps
+
+type SharedRelationshipInputProps =
+  | {
+      readonly hasMany: false
+      readonly onChange: (value: ValueWithRelation, modifyForm?: boolean) => void
+      readonly value?: null | ValueWithRelation
+    }
+  | {
+      readonly hasMany: true
+      readonly onChange: (value: ValueWithRelation[], modifyForm?: boolean) => void
+      readonly value?: null | ValueWithRelation[]
+    }
